# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/v1/public/dapp/manage/banner/list": {
            "post": {
                "description": "获取轮播图列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "dapp管理"
                ],
                "summary": "获取轮播图列表",
                "operationId": "GetManageBannerList",
                "parameters": [
                    {
                        "description": "请求体",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/protos.GetManageBannerListReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/protos.GetManageBannerListResp"
                        }
                    }
                }
            }
        },
        "/api/v1/public/dapp/manage/banner/operation": {
            "post": {
                "description": "轮播图操作管理",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "dapp管理"
                ],
                "summary": "轮播图操作管理",
                "operationId": "ManageBannerOperation",
                "parameters": [
                    {
                        "description": "请求体",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/protos.ManageBannerOperationReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/protos.ManageBannerOperationResp"
                        }
                    }
                }
            }
        },
        "/api/v1/public/dapp/manage/discover/tool/categories/list": {
            "post": {
                "description": "工具分类列表管理",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "发现页管理"
                ],
                "summary": "工具分类列表管理",
                "operationId": "GetManageDiscoverToolCategoriesList",
                "parameters": [
                    {
                        "description": "请求体",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/protos.GetManageDiscoverToolCategoriesListReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/protos.GetManageDiscoverToolCategoriesListResp"
                        }
                    }
                }
            }
        },
        "/api/v1/public/dapp/manage/discover/tool/categories/operation": {
            "post": {
                "description": "工具分类操作管理",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "发现页管理"
                ],
                "summary": "工具分类操作管理",
                "operationId": "ManageDiscoverToolCategoriesOperation",
                "parameters": [
                    {
                        "description": "请求体",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/protos.ManageDiscoverToolCategoriesOperationReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/protos.ManageDiscoverToolCategoriesOperationResp"
                        }
                    }
                }
            }
        },
        "/api/v1/public/dapp/manage/discover/tool/info/list": {
            "post": {
                "description": "工具列表管理",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "发现页管理"
                ],
                "summary": "工具列表管理",
                "operationId": "GetManageDiscoverToolInfoList",
                "parameters": [
                    {
                        "description": "请求体",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/protos.GetManageDiscoverToolInfoListReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/protos.GetManageDiscoverToolInfoListResp"
                        }
                    }
                }
            }
        },
        "/api/v1/public/dapp/manage/discover/tool/info/operation": {
            "post": {
                "description": "工具操作管理",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "发现页管理"
                ],
                "summary": "工具操作管理",
                "operationId": "ManageDiscoverToolInfoOperation",
                "parameters": [
                    {
                        "description": "请求体",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/protos.ManageDiscoverToolInfoOperationReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/protos.ManageDiscoverToolInfoOperationResp"
                        }
                    }
                }
            }
        },
        "/api/v1/public/dapp/manage/discover/toolbar/list": {
            "post": {
                "description": "发现页工具栏列表管理",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "发现页管理"
                ],
                "summary": "发现页工具栏列表管理",
                "operationId": "GetManageDiscovertoolbarList",
                "parameters": [
                    {
                        "description": "请求体",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/protos.GetManageDiscovertoolbarListReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/protos.GetManageDiscovertoolbarListResp"
                        }
                    }
                }
            }
        },
        "/api/v1/public/dapp/manage/discover/toolbar/operation": {
            "post": {
                "description": "发现页工具栏操作管理",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "发现页管理"
                ],
                "summary": "发现页工具栏操作管理",
                "operationId": "ManageDiscovertoolbarOperation",
                "parameters": [
                    {
                        "description": "请求体",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/protos.ManageDiscovertoolbarOperationReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/protos.ManageDiscovertoolbarOperationResp"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "protos.GetManageBannerListReq": {
            "type": "object",
            "properties": {
                "page": {
                    "type": "integer"
                },
                "page_size": {
                    "type": "integer"
                }
            }
        },
        "protos.GetManageBannerListResp": {
            "type": "object",
            "properties": {
                "created_time": {
                    "description": "创建时间",
                    "type": "string"
                },
                "expiration_time": {
                    "description": "过期时间",
                    "type": "string"
                },
                "id": {
                    "description": "主键ID",
                    "type": "integer"
                },
                "img_href": {
                    "description": "图片地址",
                    "type": "string"
                },
                "need_login": {
                    "description": "是否需要登录：1-不需要，2-需要",
                    "type": "integer"
                },
                "position": {
                    "description": "位置 key",
                    "type": "string"
                },
                "skip_target": {
                    "description": "跳转目标：1-H5,2-客服",
                    "type": "integer"
                },
                "skip_url": {
                    "description": "跳转URL",
                    "type": "string"
                },
                "sort": {
                    "description": "顺序",
                    "type": "integer"
                },
                "start_at": {
                    "description": "开始时间",
                    "type": "string"
                },
                "status": {
                    "description": "状态：1-上架，2-下架",
                    "type": "integer"
                },
                "title": {
                    "description": "标题",
                    "type": "string"
                },
                "updated_time": {
                    "description": "更新时间",
                    "type": "string"
                }
            }
        },
        "protos.GetManageDiscoverToolCategoriesListReq": {
            "type": "object",
            "properties": {
                "page": {
                    "type": "integer"
                },
                "page_size": {
                    "type": "integer"
                }
            }
        },
...
        "protos.ManageBannerOperationResp": {
            "type": "object"
        },
...
        "protos.ManageDiscoverToolCategoriesOperationResp": {
            "type": "object"
        },
...
        "protos.ManageDiscoverToolInfoOperationResp": {
            "type": "object"
        },
...
        "protos.ManageDiscovertoolbarOperationResp": {
            "type": "object"
        },
...
    }
}